feat: 포트폴리오 공통 컴포넌트로 분리 (DURI-184)
